requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
Werkzeug==3.0.1
insightface==0.7.3
onnxruntime-gpu==1.16.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
mysqlclient==2.2.0

# Optional but recommended
python-dotenv==1.0.0